KernelHerding

Documentation for KernelHerding.

KernelHerding.KernelHerdingIterate

KernelHerding.MeanElement

KernelHerding.NonZeroMeanElement

KernelHerding.ZeroMeanElement

Base.:*

Base.:+

Base.:-

KernelHerding.bernoulli_polynomial

KernelHerding.construct_rho

KernelHerding.kernel_evaluation_wahba

KernelHerding.mu_from_rho

KernelHerding.pad_non_zero_mean_element!

LinearAlgebra.dot

LinearAlgebra.dot

LinearAlgebra.dot

LinearAlgebra.norm

LinearAlgebra.norm

KernelHerding.KernelHerdingIterate — Type

Infinite dimensional iterate for kernel herding.

KernelHerding.MeanElement — Type

MeanElement must implement dot with a functional.

KernelHerding.NonZeroMeanElement — Type

mu =/= 0.

KernelHerding.ZeroMeanElement — Type

mu = 0.

Base.:* — Method

Multiplication for KernelHerdingIterate.

Base.:+ — Method

Addition for KernelHerdingIterate.

Base.:- — Method

Subtraction for KernelHerdingIterate.

KernelHerding.bernoulli_polynomial — Method

The degree 2 Bernoulli polynomial.

KernelHerding.construct_rho — Method

Given rho =(rhoa, rhob), returns a normalizedrho such that normalizedrho is a valid distribution.

KernelHerding.kernel_evaluation_wahba — Method

Evaluates the Wahba kernel over two real numbers.

KernelHerding.mu_from_rho — Method

Given a distribution ρ, computes μ.

KernelHerding.pad_non_zero_mean_element! — Method

The NonZeroMeanElement is represented by two vectors, cosineweights and sineweights. This function pads the shorter vector with zeros.

LinearAlgebra.dot — Method

Scalar product for two KernelHerdingIterates.

LinearAlgebra.dot — Method

Scalar product for KernelHerdingIterate with NonZeroMeanElement.

LinearAlgebra.dot — Method

Scalar product for KernelHerdingIterate with ZeroMeanElement.

LinearAlgebra.norm — Method

Norm of NonZeroMeanElement, corresponds to ||µ||.

LinearAlgebra.norm — Method

Norm of ZeroMeanElement, corresponds to ||µ||.