KernelHerding.KernelHerdingGradient — Type

The gradient < x - μ, . > is represented by x and μ.

KernelHerding.KernelHerdingIterate — Type

Kernel herding iterate.

KernelHerding.MarginalPolytopeWahba — Type

The marginal polytope of the Wahba kernel.

KernelHerding.MeanElement — Type

MeanElement μ must implement dot with a functional.

KernelHerding.NonZeroMeanElement — Type

μ =/= 0.

KernelHerding.ZeroMeanElement — Type

μ = 0.

Base.:* — Method

Multiplication for KernelHerdingIterate.

Base.:* — Method

Multiplication for KernelHerdingIterate, different order.

Base.:+ — Method

Addition for KernelHerdingIterate.

Base.:- — Method

Subtraction for KernelHerdingIterate.

FrankWolfe.compute_extreme_point — Method

Computes the extreme point in the Frank-Wolfe algorithm for kernel herding in Wahba's kernel.

KernelHerding.bernoulli_polynomial — Method

The degree-2 Bernoulli polynomial.

KernelHerding.construct_rho — Method

Given rho =(rhoa, rhob), returns a normalizedrho such that normalizedrho is a valid distribution.

KernelHerding.create_loss_function_gradient — Method

Creates the loss function and the gradient function, given a MeanElement μ, that is, 1/2 || x - μ ||_H² and < x - μ, . >, respectively.

KernelHerding.kernel_evaluation_wahba — Method

Evaluates the Wahba kernel over two real numbers.

KernelHerding.mu_from_rho — Method

Given a distribution ρ, computes μ.

KernelHerding.pad_non_zero_mean_element! — Method

The NonZeroMeanElement is represented by two vectors, cosineweights and sineweights. This function pads the shorter vector with zeros.

LinearAlgebra.dot — Method

Scalar product for KernelHerdingIterate with KernelHerdingGradient.

LinearAlgebra.dot — Method

Scalar product for two KernelHerdingIterates.

LinearAlgebra.dot — Method

Scalar product for KernelHerdingIterate with NonZeroMeanElement.

LinearAlgebra.dot — Method

Scalar product for KernelHerdingIterate with ZeroMeanElement.

LinearAlgebra.norm — Method

Norm of NonZeroMeanElement, corresponds to ||µ||.

LinearAlgebra.norm — Method

Norm of ZeroMeanElement, corresponds to ||µ||.